requirement.txt
# Django
Django==5.1.3
django-cors-headers==4.6.0

# DRF
djangorestframework==3.15.2

# swagger
drf-spectacular==0.27.2

# database
psycopg2-binary==2.9.10

# simple jwt authentication
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1

# environment
python-dotenv==1.0.1
PyYAML==6.0.2


